fix: tests & anaylitics
