+++ src/analytics/analytics.module.ts
@@ -1,11 +1,10 @@
 import { Module, forwardRef } from '@nestjs/common';
 import { ClientsModule, Transport } from '@nestjs/microservices';
 import { join } from 'path';
-import { ANALYTICS_PACKAGE_NAME } from '../analytics/balance.pb';
-import { AnalyticsController } from './analytics.controller';
-import { AnalyticsService } from './analytics.service';
-import { ANALYTICS_SERVICE_NAME } from './general.pb';
 import { UserModule } from '../user/user.module';
+import { AnalyticsController } from './analytics.controller';
+import { ANALYTICS_PACKAGE_NAME, ANALYTICS_SERVICE_NAME } from './analytics.pb';
+import { AnalyticsService } from './analytics.service';
 
 @Module({
   imports: [
@@ -17,7 +16,7 @@ import { UserModule } from '../user/user.module';
           url: process.env.ANALYTICS_SERVICE_URL,
           package: ANALYTICS_PACKAGE_NAME,
           protoPath: join(
-            'node_modules/@fidecly/grpc-proto/proto/analytics/general.proto',
+            'node_modules/@fidecly/grpc-proto/proto/analytics/analytics.proto',
           ),
         },
       },
