+++ src/balance/balance.module.ts
@@ -1,17 +1,17 @@
 import { forwardRef, Module } from '@nestjs/common';
+import { ClientsModule, Transport } from '@nestjs/microservices';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { CardModule } from '../card/card.module';
-import { PromotionModule } from '../promotion/promotion.module';
-import { BalanceController } from './balance.controller';
-import { Balance } from './balance.entity';
-import { BalanceService } from './balance.service';
-import { UserModule } from '../user/user.module';
+import { join } from 'path';
 import {
   ANALYTICS_PACKAGE_NAME,
   BALANCE_SERVICE_NAME,
-} from '../analytics/balance.pb';
-import { ClientsModule, Transport } from '@nestjs/microservices';
-import { join } from 'path';
+} from '../analytics/analytics.pb';
+import { CardModule } from '../card/card.module';
+import { PromotionModule } from '../promotion/promotion.module';
+import { UserModule } from '../user/user.module';
+import { BalanceController } from './balance.controller';
+import { Balance } from './balance.entity';
+import { BalanceService } from './balance.service';
 
 @Module({
   imports: [
@@ -23,7 +23,7 @@ import { join } from 'path';
           url: process.env.ANALYTICS_SERVICE_URL,
           package: ANALYTICS_PACKAGE_NAME,
           protoPath: join(
-            'node_modules/@fidecly/grpc-proto/proto/analytics/balance.proto',
+            'node_modules/@fidecly/grpc-proto/proto/analytics/analytics.proto',
           ),
         },
       },
