+++ src/campaign/campaign.pb.ts
@@ -1,8 +1,8 @@
 /* eslint-disable */
-import { GrpcMethod, GrpcStreamMethod } from '@nestjs/microservices';
-import { Observable } from 'rxjs';
+import { GrpcMethod, GrpcStreamMethod } from "@nestjs/microservices";
+import { Observable } from "rxjs";
 
-export const protobufPackage = 'campaign';
+export const protobufPackage = "campaign";
 
 export interface SendRequest {
   senderEmail: string;
@@ -18,45 +18,29 @@ export interface SendResponse {
   errors: string[];
 }
 
-export const CAMPAIGN_PACKAGE_NAME = 'campaign';
+export const CAMPAIGN_PACKAGE_NAME = "campaign";
 
 export interface CampaignServiceClient {
   send(request: SendRequest): Observable<SendResponse>;
 }
 
 export interface CampaignServiceController {
-  send(
-    request: SendRequest,
-  ): Promise<SendResponse> | Observable<SendResponse> | SendResponse;
+  send(request: SendRequest): Promise<SendResponse> | Observable<SendResponse> | SendResponse;
 }
 
 export function CampaignServiceControllerMethods() {
   return function (constructor: Function) {
-    const grpcMethods: string[] = ['send'];
+    const grpcMethods: string[] = ["send"];
     for (const method of grpcMethods) {
-      const descriptor: any = Reflect.getOwnPropertyDescriptor(
-        constructor.prototype,
-        method,
-      );
-      GrpcMethod('CampaignService', method)(
-        constructor.prototype[method],
-        method,
-        descriptor,
-      );
+      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
+      GrpcMethod("CampaignService", method)(constructor.prototype[method], method, descriptor);
     }
     const grpcStreamMethods: string[] = [];
     for (const method of grpcStreamMethods) {
-      const descriptor: any = Reflect.getOwnPropertyDescriptor(
-        constructor.prototype,
-        method,
-      );
-      GrpcStreamMethod('CampaignService', method)(
-        constructor.prototype[method],
-        method,
-        descriptor,
-      );
+      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
+      GrpcStreamMethod("CampaignService", method)(constructor.prototype[method], method, descriptor);
     }
   };
 }
 
-export const CAMPAIGN_SERVICE_NAME = 'CampaignService';
+export const CAMPAIGN_SERVICE_NAME = "CampaignService";
