+++ src/card/card.module.ts
@@ -1,16 +1,16 @@
 import { forwardRef, Module } from '@nestjs/common';
+import { ClientsModule, Transport } from '@nestjs/microservices';
 import { TypeOrmModule } from '@nestjs/typeorm';
+import { join } from 'path';
+import {
+  ANALYTICS_PACKAGE_NAME,
+  CARD_SERVICE_NAME,
+} from '../analytics/analytics.pb';
 import { ShopModule } from '../shop/shop.module';
 import { UserModule } from '../user/user.module';
 import { CardController } from './card.controller';
 import { Card } from './card.entity';
 import { CardService } from './card.service';
-import {
-  ANALYTICS_PACKAGE_NAME,
-  CARD_SERVICE_NAME,
-} from '../analytics/card.pb';
-import { ClientsModule, Transport } from '@nestjs/microservices';
-import { join } from 'path';
 
 @Module({
   imports: [
@@ -22,7 +22,7 @@ import { join } from 'path';
           url: process.env.ANALYTICS_SERVICE_URL,
           package: ANALYTICS_PACKAGE_NAME,
           protoPath: join(
-            'node_modules/@fidecly/grpc-proto/proto/analytics/card.proto',
+            'node_modules/@fidecly/grpc-proto/proto/analytics/analytics.proto',
           ),
         },
       },
