+++ src/promotion/promotion.service.ts
@@ -1,13 +1,14 @@
 import { Inject, Injectable } from '@nestjs/common';
+import { ClientGrpc } from '@nestjs/microservices';
 import { InjectRepository } from '@nestjs/typeorm';
+import { firstValueFrom } from 'rxjs';
 import { Repository, UpdateResult } from 'typeorm';
+import {
+  PROMOTION_SERVICE_NAME,
+  PromotionServiceClient,
+} from '../analytics/analytics.pb';
 import { CreatePromotionDto, UpdatePromotionDto } from './promotion.dto';
 import { Promotion } from './promotion.entity';
-import { ClientGrpc } from '@nestjs/microservices';
-import {
-  PromotionServiceClient,
-  PROMOTION_SERVICE_NAME,
-} from '../analytics/promotion.pb';
 
 @Injectable()
 export class PromotionService {
@@ -63,10 +64,12 @@ export class PromotionService {
   // ANALYTICS
 
   sendToAnalytics(promotion: Promotion) {
-    this.analyticsService.send({
-      ...promotion,
-      startAt: new Date(promotion.startAt).toISOString(),
-      endAt: new Date(promotion.endAt).toISOString(),
-    });
+    return firstValueFrom(
+      this.analyticsService.send({
+        ...promotion,
+        startAt: new Date(promotion.startAt).toISOString(),
+        endAt: new Date(promotion.endAt).toISOString(),
+      }),
+    );
   }
 }
