+++ src/shop/shop.module.ts
@@ -1,18 +1,18 @@
 import { forwardRef, Module } from '@nestjs/common';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { PromotionModule } from '../promotion/promotion.module';
-import { CardModule } from '../card/card.module';
-import { ShopController } from './shop.controller';
-import { Shop } from './shop.entity';
-import { ShopService } from './shop.service';
-import { UserModule } from '../user/user.module';
-import { CampaignModule } from '../campaign/campaign.module';
 import { ClientsModule, Transport } from '@nestjs/microservices';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { join } from 'path';
 import {
   ANALYTICS_PACKAGE_NAME,
   SHOP_SERVICE_NAME,
-} from '../analytics/shop.pb';
-import { join } from 'path';
+} from '../analytics/analytics.pb';
+import { CampaignModule } from '../campaign/campaign.module';
+import { CardModule } from '../card/card.module';
+import { PromotionModule } from '../promotion/promotion.module';
+import { UserModule } from '../user/user.module';
+import { ShopController } from './shop.controller';
+import { Shop } from './shop.entity';
+import { ShopService } from './shop.service';
 
 @Module({
   imports: [
@@ -24,7 +24,7 @@ import { join } from 'path';
           url: process.env.ANALYTICS_SERVICE_URL,
           package: ANALYTICS_PACKAGE_NAME,
           protoPath: join(
-            'node_modules/@fidecly/grpc-proto/proto/analytics/shop.proto',
+            'node_modules/@fidecly/grpc-proto/proto/analytics/analytics.proto',
           ),
         },
       },
