+++ src/shop/shop.service.ts
@@ -1,10 +1,14 @@
 import { Inject, Injectable } from '@nestjs/common';
+import { ClientGrpc } from '@nestjs/microservices';
 import { InjectRepository } from '@nestjs/typeorm';
+import { firstValueFrom } from 'rxjs';
 import { Repository, UpdateResult } from 'typeorm';
+import {
+  SHOP_SERVICE_NAME,
+  ShopServiceClient,
+} from '../analytics/analytics.pb';
 import { CreateShopDto, UpdateShopDto } from './shop.dto';
 import { Shop } from './shop.entity';
-import { SHOP_SERVICE_NAME, ShopServiceClient } from '../analytics/shop.pb';
-import { ClientGrpc } from '@nestjs/microservices';
 
 @Injectable()
 export class ShopService {
@@ -177,6 +181,6 @@ export class ShopService {
   // ANALYTICS
 
   sendToAnalytics(shop: Shop) {
-    this.analyticsService.send(shop);
+    return firstValueFrom(this.analyticsService.send(shop));
   }
 }
